+++ apps/property-management-api/src/app.module.ts
@@ -2,16 +2,16 @@ import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { HttpModule } from '@nestjs/axios';
 import { loadConfig } from './config';
-// TODO: import { RequestsModule } from './requests/requests.module';
+import { MaintenanceRequestModule } from './maintenance-request/maintenance-request.module';
 
 @Module({
     imports: [
         ConfigModule.forRoot({
             isGlobal: true,
-            load: [loadConfig],
+            load: [loadConfig]
         }),
         HttpModule,
-        // TODO: RequestsModule
+        MaintenanceRequestModule,
     ]
 })
 export class AppModule {}