+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2025-06-02
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): `MaintenanceRequestModule` containing two endpoints: `GET /requests?priority` and `POST /requests` to add a new one.
+- [Marek Kurańda](https://github.com/mjkuranda): `DynamodbModule` to integrate DynamoDB.
+- [Marek Kurańda](https://github.com/mjkuranda): Thresholds for defining priority levels.
+
 ## [1.1.0] - 2025-05-31
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Keyword config for all words or tokens to define priority level.